@@ -1,16 +1,15 @@
 // Importing various dependancies to be used later on in the project.
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import './style.css'
 import App from './App'
 import { ErrorBoundary } from './components/ErrorBoundary';
 
 // If successful - the App page is loaded and if not - the Error Boundary page gets loaded.
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <ErrorBoundary>
       <App />
     </ErrorBoundary>
-  </React.StrictMode>,
-  document.getElementById('root')
-)
+  </React.StrictMode>
+);
